[1822MX] Fix advanced tile lay and remove town privates
